Complete Phase 20 dependency consolidation
Adopt base64, clap (derive), directories, and thiserror; migrate the four
hand-rolled areas with byte-identical behavior and parity tests. Widen the
graphics z-index to the full i32 range and add the real-Kitty animated GIF
and composition harness. Plan Phases 17-19 and Workstream 8 in the roadmap
and reconcile DEPENDENCIES.md with Cargo.toml.

+++ src/api.rs
@@ -427,7 +427,7 @@ pub struct GraphicsDto {
     pub generation: u64,
     pub format: u8,
     pub area: RectDto,
-    pub z_index: i16,
+    pub z_index: i32,
 }
 
 impl From<&crate::graphics::GraphicsSubmission> for GraphicsDto {
